parser/requirements.txt
selenium==4.8.2
urllib3==1.26.15
certifi==2022.12.7
python-dotenv==1.0.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9

